Don't return the in-flight SQL handler
Seems we can now get a failure on jammy where the commit/rollback
has *already* occured in some cases, causing test failures.

This seems to be a result of the pattern of returning the in-flight
db operation, instead of returning the result later.

Since we should only *really* be doing so where we explicitly
know we need to do so, I've gone ahead and changed the rest of
the returns, since we recently had to visit some of them for
the same basic reason.

Change-Id: I3c2033b05f8005891d96c3ca141a80838a8699a4